feat: Add build script on maven and move files from practice to grep
