+++ core_java/grep/src/main/java/ca/jrvs/apps/grep/JavaGrep.java
@@ -1,4 +1,4 @@
-package ca.jrvs.apps.practice;
+package ca.jrvs.apps.grep;
 
 import java.io.File;
 import java.io.IOException;
