+++ core_java/grep/src/main/java/ca/jrvs/apps/grep/JavaGrepLambdaImp.java
@@ -1,4 +1,4 @@
-package ca.jrvs.apps.practice;
+package ca.jrvs.apps.grep;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
